a7xelte: disable fingerprint sensor when screen is off
When the screen is off, the fingerprint sensor scans every second for
a finger which causes unwanted battery drain. Disabling the sensor
when the device is sleeping extends the standby time by a factor of
about 2-3x.
As the fingerprint reader is also the power button, this doesn't
hinder its usage.

Change-Id: I347f5b060723ca663c683786ea10c3efe8a18782
Ref: BUGBASH-1422